+++ scripts/case/transcribe1.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env sh
+
+# Run this from the root of cyllama
+
+uv run cyllama transcribe \
+	-f tests/samples/jfk.wav \
+	-m models/ggml-base.en.bin